hf-custom-handler/requirements.txt
diffusers==0.27.2
transformers==4.38.0
accelerate==0.27.0
safetensors==0.4.2
torch==2.1.2
torchvision==0.16.2
Pillow==10.2.0



